refactor: Drop FlowUtils
This is a hangover from the UBA. Relocate the function to SpokeUtils
instead. This will make it easier to group functionality into evm/xvm.
